Temporarily remove flyout caching.
This is really problematic for the keyboard nav plugin. The plugin assumes it
can access the flyout's workspace at initialization time.

There's a largely pointless delegation pattern here that we can remove if we
keep this but likely we need to do something to address performance if it's
still an issue. Hoping this is good enough for user testing at least.