Improve block comment keyboard navigation
Workaround for the block comment part of
https://github.com/microsoft/pxt-microbit/issues/6404

Block comments aren't yet navigable as part of the normal sequence and won't be
for the initial 2025 release. As a workaround we now:
- focus the comment when opening it
- allow escape to close the comment moving focus to the block

This allows users to work with comments in a toggle on/off fashion which is all
that's viable until they're in the normal navigation flow. When that happens,
focus behaviour might need to change to allow access to the block comment
toolbar in a manner similar to the Blockly work on workspace comments.

The preventDefault calls are needed to prevent loss of focus to the body when
clicking the delete/collapse icons.